Python_programming_assignment/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
supabase==2.7.2
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pandas==2.2.3
python-multipart==0.0.12
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.25.0
